+++ Makefile
@@ -5,3 +5,4 @@ all:
 	# g++ -std=c++0x -Wall $(INC) ATMclient.cpp -lcryptopp -lpthread -o ATMclient -g
 	# g++ -Wall $(INC) ATMproxy.cpp -o ATMproxy -lpthread -g
 	g++ $(INC) ATMserver.cpp -lcryptopp -lpthread -o ATMserver
+	g++ $(INC) ATMproxy.cpp -lcryptopp -lpthread -o ATMproxy
